feat(http): extend /v1/stats and split handlers per endpoint (dashboard step 2)
Two changes:

1. Extract the inline handlers from router.rs into per-endpoint files
   under crates/ctxd-http/src/handlers/. Pure refactor: each handler
   keeps the same auth model, error mapping, and behavior. Shared
   helpers (bearer_from_headers, require_admin) live in handlers/mod.rs.
   router.rs is now ~60 lines: AppState, build_router, route table.

2. Extend /v1/stats from { subject_count } to the v0.4 shape:
     event_count, subject_count, peer_count, pending_approval_count,
     vector_embedding_count, uptime_seconds, version
   The shape is purely additive — older clients reading subject_count
   keep working. Each counter is best-effort (warn-and-zero on store
   failure) because the dashboard would rather render `0` than 500
   when one counter hiccups.

AppState gains a `start_time: Instant` captured at build_router()
time; uptime_seconds is derived from `start_time.elapsed()`. There
is no std::process::start_time, and the gap between binary start and
router build is measured in milliseconds — close enough for an
operator-facing counter.

Verified: cargo test -p ctxd-http (19 tests; 2 new in stats_shape +
17 existing) all pass; cargo build --workspace clean; smoke against
`ctxd serve --bind 127.0.0.1:17777` shows the full extended JSON
shape live (event_count, peer_count, pending_approval_count,
subject_count, uptime_seconds, vector_embedding_count, version).
